Add files via upload
The goal of this project is to build a model that predicts the rating of a movie based on features like genre, director, and actors. Using regression techniques, the project aims to analyze historical movie data and develop a model that accurately estimates the rating given to a movie by users or critics. This project enables you to explore data analysis, preprocessing, feature engineering, and machine learning modeling techniques. It provides insights into the factors that influence movie ratings and allows you to build a model that can estimate the ratings of movies accurately.