Add files via upload
The goal of this project is to build a machine learning model to identify fraudulent credit card transactions. The project involves preprocessing and normalizing the transaction data, handling class imbalance issues, and splitting the dataset into training and testing sets. A classification algorithm, such as logistic regression or random forests, is trained to classify transactions as fraudulent or genuine. The model's performance is evaluated using metrics like precision, recall, and F1-score, with techniques like oversampling or undersampling considered for improving results.